% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anak_tidak_sekolah.R
\docType{data}
\name{anak_tidak_sekolah}
\alias{anak_tidak_sekolah}
\title{Dataset Angka Anak Tidak Sekolah Menurut Jenjang Pendidikan dan Jenis Kelamin (2019-2023)}
\format{
Objek time series dengan 5 observasi yang dirinci berdasarkan jenjang pendidikan dan
jenis kelamin.
Nilai mulai dari tahun 2019 hingga 2023.
}
\source{
Badan Pusat Statistik (BPS)
}
\usage{
anak_tidak_sekolah
}
\description{
Data ini berisi Angka Anak Tidak Sekolah Menurut Jenjang Pendidikan dan Jenis Kelamin
Indonesia yang dirilis oleh Badan Pusat Statistik (BPS), dalam satuan persen (%).
}
\examples{
tidak_sekolah_sd <- anak_tidak_sekolah \%>\% select(contains(c("tahun", "sd")))
plot(
  tidak_sekolah_sd, main = "Angka Anak Tidak Sekolah Indonesia - Jenjang SD (2019-2023)",
  ylab = "Persen (\%)", xlab = "Tahun"
)
}
\keyword{datasets}
