% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indeks_ketimpangan_gender.R
\docType{data}
\name{indeks_ketimpangan_gender}
\alias{indeks_ketimpangan_gender}
\title{Dataset Indeks Ketimpangan Gender (2018-2024)}
\format{
Objek time series dengan 7 observasi dan frekuensi tahunan (\code{frequency = 1}).
Nilai mulai dari Tahun 2018 hingga Tahun 2024.
}
\source{
Badan Pusat Statistik (BPS)
}
\usage{
indeks_ketimpangan_gender
}
\description{
Data ini berisi Indeks Ketimpangan Gender (IKG) Indonesia yang dirilis tahunan
oleh Badan Pusat Statistik (BPS), dalam satuan (indeks). IKG adalah pengganti ukuran Indeks Pemberdayaan Gender
}
\examples{
indeks_ketimpangan_gender <- indeks_ketimpangan_gender
plot(
  indeks_ketimpangan_gender, main = "Indeks Ketimpangan Gender Indonesia (2018-2024)",
  ylab = "Persen (\%)", xlab = "Tahun"
)
}
\keyword{datasets}
