% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indeks_pembangunan_gender.R
\docType{data}
\name{indeks_pembangunan_gender}
\alias{indeks_pembangunan_gender}
\title{Dataset Indeks Pembangunan Gender (2020-2024) dengan Metode Terbaru}
\format{
Objek time series dengan 5 observasi dan frekuensi tahunan (`frequency = 1`).
Nilai mulai dari Tahun 2020 hingga Tahun 2024.
}
\source{
Badan Pusat Statistik (BPS)
}
\usage{
indeks_pembangunan_gender
}
\description{
Data ini berisi Indeks Pembangunan Gender (IPG) Indonesia yang dirilis tahunan
oleh Badan Pusat Statistik (BPS), dalam satuan tahun. IPG ini menggunakan metode baru
yaitu data Umur Harapan Hidup hasil SP2020 LF
}
\examples{
ipg <- indeks_pembangunan_gender
plot(
  ipg, main = "Indeks Pembangunan Gender Indonesia (2010-2024)",
  ylab = "Tahun", xlab = "Tahun"
)
}
\keyword{datasets}
