% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indeks_pembangunan_manusia.R
\docType{data}
\name{indeks_pembangunan_manusia}
\alias{indeks_pembangunan_manusia}
\title{Dataset Indeks Pembangunan Manusia (2010–2024) dengan Metode Terbaru}
\format{
Objek time series dengan 15 observasi dan frekuensi tahunan (\code{frequency = 1}).
Nilai mulai dari Tahun 2010 hingga Tahun 2024.
}
\source{
Badan Pusat Statistik (BPS)
}
\usage{
indeks_pembangunan_manusia
}
\description{
Data ini berisi Indeks Pembangunan Manusia (IPM) Indonesia yang dirilis tahunan
oleh Badan Pusat Statistik (BPS), dalam satuan persen (\%). IPM ini menggunakan metode baru dan telah disesuaikan dengan angka Umur Harapan Hidup Hasil Long Form SP2020
}
\examples{
indeks_pembangunan_manusia <- indeks_pembangunan_manusia
plot(
  indeks_pembangunan_manusia, main = "Indeks Pembangunan Manusia Indonesia (2010–2024)",
  ylab = "Persen (\%)", xlab = "Tahun"
)
}
\keyword{datasets}
