% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penduduk_bekerja_pengangguran.R
\docType{data}
\name{penduduk_bekerja_pengangguran}
\alias{penduduk_bekerja_pengangguran}
\title{Dataset Jumlah dan Persentase Penduduk Bekerja dan Pengangguran Indonesia
(Februari 2011–Februari 2025)}
\format{
Sebuah data frame dengan 15 observasi dan 5 variabel:
\describe{
\item{tahun}{Tahun pengamatan (integer).}
\item{jml_kerja}{Jumlah penduduk bekerja (ribu orang).}
\item{persentase_kerja}{Persentase penduduk bekerja (\%).}
\item{jml_pengangguran}{Jumlah pengangguran (ribu orang).}
\item{persentase_pengangguran}{Persentase pengangguran (\%).}
}
}
\source{
Badan Pusat Statistik (BPS)
}
\usage{
penduduk_bekerja_pengangguran
}
\description{
Dataset ini memuat jumlah dan persentase penduduk bekerja serta
pengangguran di Indonesia berdasarkan hasil Sakernas bulan Februari
periode 2011–2025.
}
\examples{
data(penduduk_bekerja_pengangguran)

if (interactive()) {
  plot(
    penduduk_bekerja_pengangguran$tahun,
    penduduk_bekerja_pengangguran$persentase_pengangguran,
    type = "l",
    xlab = "Tahun",
    ylab = "Persentase (\%)",
    main = "Persentase Pengangguran Tahunan"
  )
}
}
\keyword{datasets}
