% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rata2_lama_sekolah.R
\docType{data}
\name{rata2_lama_sekolah}
\alias{rata2_lama_sekolah}
\title{Dataset Rata-rata Lama Sekolah (2010-2024) dengan Metode Terbaru}
\format{
Objek time series dengan 15 observasi dan frekuensi tahunan (`frequency = 1`).
Nilai mulai dari Tahun 2010 hingga Tahun 2024.
}
\source{
Badan Pusat Statistik (BPS)
}
\usage{
rata2_lama_sekolah
}
\description{
Data ini berisi Rata-rata Lama Sekolah (RLS) Indonesia yang dirilis tahunan
oleh Badan Pusat Statistik (BPS), dalam satuan tahun. RLS ini menggunakan metode baru dan telah disesuaikan dengan angka Umur Harapan Hidup Hasil Long Form SP2020
}
\examples{
rata2_lama_sekolah <- rata2_lama_sekolah
plot(
  rata2_lama_sekolah, main = "Rata-rata Lama Sekolah Indonesia (2010-2024)",
  ylab = "Tahun", xlab = "Tahun"
)
}
\keyword{datasets}
