% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_xl.R
\name{save_xl}
\alias{save_xl}
\title{Menyimpan dataset terpilih}
\usage{
save_xl(dataset)
}
\value{
File Excel dari dataset dan informasi direktori dari file tersimpan.
}
\description{
Fungsi ini berguna untuk Menyimpan dataset terpilih di paket bpsuseR
dalam format Excel.
}
\examples{
save_xl(tingkat_pengangguran_terbuka)

}
