% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umur_harapan_hidup.R
\docType{data}
\name{umur_harapan_hidup}
\alias{umur_harapan_hidup}
\title{Dataset Umur Harapan Hidup (2010-2024) dengan Metode Terbaru}
\format{
Objek time series dengan 15 observasi dan frekuensi tahunan (`frequency = 1`).
Nilai mulai dari Tahun 2010 hingga Tahun 2024.
}
\source{
Badan Pusat Statistik (BPS)
}
\usage{
umur_harapan_hidup
}
\description{
Data ini berisi Umur Harapan Hidup (UHH) Indonesia yang dirilis tahunan
oleh Badan Pusat Statistik (BPS), dalam satuan tahun. UHH ini menggunakan metode baru
yang merupakan salah satu indikator dari Indeks Pembangunan Manusia (IPM)
}
\examples{
uhh <- umur_harapan_hidup
plot(
  uhh, main = "Umur Harapan Hidup Indonesia (2010-2024)",
  ylab = "Tahun", xlab = "Tahun"
)
}
\keyword{datasets}
