% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_bps_data.R
\name{list_bps_data}
\alias{list_bps_data}
\title{Menampilkan daftar dataset yang tersedia dalam package bpsuseR}
\usage{
list_bps_data()
}
\description{
Fungsi ini menampilkan seluruh dataset yang terdapat dalam package bpsuseR.
}
