% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_adhk_tahunan.R
\docType{data}
\name{pdb_adhk_tahunan}
\alias{pdb_adhk_tahunan}
\title{Produk Domestik Bruto (ADHK) Tahunan Indonesia, 2010–2024}
\format{
Objek time series (\code{ts}) frekuensi tahunan (2010–2024).
}
\source{
Badan Pusat Statistik (BPS)
}
\usage{
pdb_adhk_tahunan
}
\description{
Dataset ini berisi nilai Produk Domestik Bruto (PDB) atas dasar harga konstan (ADHK)
dari tahun 2010 hingga 2024 (dalam miliar rupiah) dalam bentuk deret waktu tahunan.
}
\examples{
data(pdb_adhk_tahunan)
plot(pdb_adhk_tahunan, main = "PDB ADHK Tahunan Indonesia (2010–2024)",
     ylab = "Miliar Rupiah", xlab = "Tahun")
}
\keyword{datasets}
