% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penduduk_bekerja_pengangguran.R
\docType{data}
\name{penduduk_bekerja_pengangguran}
\alias{penduduk_bekerja_pengangguran}
\title{Dataset Jumlah dan Persentase Penduduk Bekerja dan Pengangguran Indonesia
(Februari 2011 - Februari 2025)}
\format{
Objek time series dengan 15 observasi dan 4 variabel (jml_kerja, persentase kerja,
jml_pengangguran, persentase_penngangguran).
Nilai mulai dari Februari 2011 hingga Februari 2025.
}
\source{
Badan Pusat Statistik (BPS)
}
\usage{
penduduk_bekerja_pengangguran
}
\description{
Data ini berisi Jumlah dan Persentase Penduduk Bekerja dan Pengangguran Indonesia
di bulan Februari (2011 - 2025) yang dirilis oleh Badan Pusat Statistik (BPS).
}
\keyword{datasets}
