% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_xl.R
\name{save_xl}
\alias{save_xl}
\title{Menyimpan dataset terpilih}
\usage{
save_xl(dataset, path = getwd())
}
\arguments{
\item{dataset}{Objek data frame atau objek yang dapat
dikonversi ke data frame.}
}
\value{
File CSV yang tersimpan di direktori tujuan dan pesan
konfirmasi lokasi file.
}
\description{
Fungsi ini digunakan untuk menyimpan dataset yang tersedia
di paket \pkg{bpsuseR} ke dalam berkas CSV menggunakan fungsi
bawaan R tanpa dependensi tambahan.
}
\examples{
data(tingkat_pengangguran_terbuka)

if (interactive()) {
  save_xl(
    tingkat_pengangguran_terbuka,
    path = tempdir()
  )
}

}
