% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indeks_pembangunan_manusia.R,
%   R/tingkat_pengangguran_terbuka.R
\docType{data}
\name{tingkat_pengangguran_terbuka}
\alias{tingkat_pengangguran_terbuka}
\title{Dataset Indeks Pembangunan Manusia (2010–2024) dengan Metode Terbaru}
\format{
Objek time series dengan 15 observasi dan frekuensi tahunan (\code{frequency = 1}).
Nilai mulai dari Tahun 2010 hingga Tahun 2024.

Objek time series dengan 41 observasi dan frekuensi dua kali per tahun (\code{frequency = 2}).
Nilai mulai dari Februari 2005 hingga Februari 2025.
}
\source{
Badan Pusat Statistik (BPS)

Badan Pusat Statistik (BPS)
}
\usage{
tingkat_pengangguran_terbuka

tingkat_pengangguran_terbuka
}
\description{
Data ini berisi Indeks Pembangunan Manusia (IPM) Indonesia yang dirilis tahunan
oleh Badan Pusat Statistik (BPS), dalam satuan persen (\%). IPM ini menggunakan metode baru dan telah disesuaikan dengan angka Umur Harapan Hidup Hasil Long Form SP2020

Data ini berisi Tingkat Pengangguran Terbuka (TPT) Indonesia yang dirilis dua kali per tahun (Februari dan Agustus)
oleh Badan Pusat Statistik (BPS), dalam satuan persen (\%).
}
\examples{
indeks_pembangunan_manusia <- indeks_pembangunan_manusia
plot(
  indeks_pembangunan_manusia, main = "Indeks Pembangunan Manusia Indonesia (2010–2024)",
  ylab = "Persen (\%)", xlab = "Tahun"
)
tingkat_pengangguran_terbuka <- tingkat_pengangguran_terbuka
plot(
  tingkat_pengangguran_terbuka, main = "Tingkat Pengangguran Terbuka Indonesia (2005–2025)",
  ylab = "Persen (\%)", xlab = "Tahun"
)
}
\keyword{datasets}
