% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tingkat_pengangguran_terbuka.R
\docType{data}
\name{tingkat_pengangguran_terbuka}
\alias{tingkat_pengangguran_terbuka}
\title{Dataset Tingkat Pengangguran Terbuka (2005-2025)}
\format{
Objek data frame dengan 42 observasi dan 3 variabel:
}
\source{
Badan Pusat Statistik (BPS)
}
\usage{
tingkat_pengangguran_terbuka
}
\description{
Data ini berisi Tingkat Pengangguran Terbuka rilis 2 kali per tahunnya
dalam persen
}
\examples{
data("tingkat_pengangguran_terbuka")
plot(tingkat_pengangguran_terbuka$Tahun, tingkat_pengangguran_terbuka$TPT, main = "Tingkat Pengangguran Terbuka Indonesia (Februari dan Agustus) 2005–2025 (persen)",
     ylab = "persen", xlab = "Tahun")
}
\keyword{datasets}
