% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usia_muda_tidak_sekolah_kerja_pelatihan.R
\docType{data}
\name{usia_muda_tidak_sekolah_kerja_pelatihan}
\alias{usia_muda_tidak_sekolah_kerja_pelatihan}
\title{Dataset Usia Muda (15-24 Tahun) Yang Sedang Tidak Sekolah, Bekerja Atau
Mengikuti Pelatihan (2015-2024)}
\format{
Objek time series dengan 10 observasi dan satu variabel berupa persentase.
Nilai mulai dari 2015 hingga 2024.
}
\source{
Badan Pusat Statistik (BPS)
}
\usage{
usia_muda_tidak_sekolah_kerja_pelatihan
}
\description{
Data ini berisi Persentase Usia Muda (15-24 Tahun) Yang Sedang Tidak Sekolah,
Bekerja Atau Mengikuti Pelatihan di Indonesia yang dirilis sekali per tahun
oleh Badan Pusat Statistik (BPS), dalam satuan persen (%).
}
\examples{
dt <- usia_muda_tidak_sekolah_kerja_pelatihan
plot(
  dt, main = "Usia Muda Yang Sedang Tidak Sekolah, Bekerja Atau Mengikuti Pelatihan
  di Indonesia (2015-2024)",
  ylab = "Persen (\%)", xlab = "Tahun"
)
}
\keyword{datasets}
